Forecast moved to class
